Move extract linkage code to the new "linkage.py" file
